+++ lib/sdui.dart
@@ -7,6 +7,7 @@ export 'src/container.dart';
 export 'src/dialog.dart';
 export 'src/divider.dart';
 export 'src/dropdown.dart';
+export 'src/error.dart';
 export 'src/expanded.dart';
 export 'src/flexible.dart';
 export 'src/form.dart';