+++ lib/src/loading.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
+typedef LoadingStateBuilder = Widget Function(BuildContext context);
+
 /// Loading state
 // ignore: prefer_function_declarations_over_variables
-var sduiLoadingState = (BuildContext context) =>  const CircularProgressIndicator();
+LoadingStateBuilder sduiLoadingState = (context) =>  const CircularProgressIndicator();